pyproject.toml
[project]
name = "dcc-poc"
version = "0.1.0"
description = "Dataspace Compliance Copilot PoC"
authors = [{name = "DCC Team"}]
requires-python = ">=3.11"
dependencies = [
    "streamlit>=1.36",
    "pyyaml>=6.0",
    "jsonschema>=4.22",
    "jsonpatch>=1.33",
    "pytest>=8.2",
    "requests>=2.31",
    "openai>=1.35",
    "python-dotenv>=1.0"
]

[project.optional-dependencies]
app = ["streamlit"]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["pipeline"]

[tool.pytest.ini_options]
addopts = "-q"
pythonpath = ["."]
